requirements.txt
flask
scikit-learn
opencv-python
numpy
pandas
Pillow
joblib
